fix wrong type on BPHour on ajax.php

+++ sections/ajax/info.php
@@ -34,7 +34,7 @@ json_print("success", [
         'ratio' => (float)$Ratio,
         'requiredratio' => (float)$LoggedUser['RequiredRatio'],
         'bonusPoints' => (int)$LoggedUser['BonusPoints'],
-        'bonusPointsPerHour' => number_format($LoggedUser['BonusPointsPerHour'], 2),
+        'bonusPointsPerHour' => (float)number_format($LoggedUser['BonusPointsPerHour'], 2),
         'class' => $ClassLevels[$LoggedUser['Class']]['Name']
     ]
 ]);
